requirements.txt
openai-agents==0.0.9
python-dotenv==1.0.0
pydantic>=2.10.0
typing-extensions>=4.12.2
